Rearranged generic unit tests
