+++ src/Serialize.Linq.Tests/IssuesGeneric/Issue31.cs
@@ -5,11 +5,11 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Serialize.Linq.Interfaces;
 using Serialize.Linq.Serializers;
 
-namespace Serialize.Linq.Tests.Issues
+namespace Serialize.Linq.Tests.IssuesGeneric
 {
     // https://github.com/esskar/Serialize.Linq/issues/31
     [TestClass]
-    public class Issue31Generic
+    public class Issue31
     {
         [TestMethod]
         public void SerializeLambdaWithEnumTest()
