+++ src/Serialize.Linq.Tests/IssuesGeneric/Issue35.cs
@@ -1,26 +1,36 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Linq.Expressions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Serialize.Linq.Interfaces;
 using Serialize.Linq.Serializers;
 using Serialize.Linq.Tests.Internals;
 
-namespace Serialize.Linq.Tests.Issues
+namespace Serialize.Linq.Tests.IssuesGeneric
 {
+    // https://github.com/esskar/Serialize.Linq/issues/35
     [TestClass]
-    public class Issue37Generic
+    public class Issue35
     {
         [TestMethod]
-        public void DynamicsTests()
+        public void LetExpressionTests()
         {
             var expressions = new List<Expression>();
 
-            Expression<Func<Item, dynamic>> objectExp = item => new {item.Name, item.ProductId};
-            Expression<Func<string, dynamic>> stringExp = str => new { Text = str };
+            Expression<Func<IEnumerable<int>, IEnumerable<int>>> intExpr = c =>
+                from x in c
+                let test = 8
+                where x == test
+                select x;
+            expressions.Add(intExpr);
 
-            expressions.Add(objectExp);
-            expressions.Add(stringExp);
+            Expression<Func<IEnumerable<string>, IEnumerable<string>>> strExpr = c =>
+                from x in c
+                let test = "bar"
+                where x == test
+                select x;
+            expressions.Add(strExpr);            
 
             foreach (var textSerializer in new ITextSerializer[] { new JsonSerializer(), new XmlSerializer() })
             {
@@ -33,13 +43,6 @@ namespace Serialize.Linq.Tests.Issues
                     ExpressionAssert.AreEqual(expected, actual);
                 }
             }
-        }
-
-        public class Item
-        {
-            public string Name { get; set; }
-
-            public string ProductId { get; set; }
         }
     }
 }
