+++ src/Serialize.Linq.Tests/IssuesGeneric/Issue56.cs
@@ -5,13 +5,13 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Serialize.Linq.Internals;
 using Serialize.Linq.Serializers;
 
-namespace Serialize.Linq.Tests.Issues
+namespace Serialize.Linq.Tests.IssuesGeneric
 {
     /// <summary>
     /// https://github.com/esskar/Serialize.Linq/issues/56
     /// </summary>
     [TestClass]
-    public class Issue56Generic
+    public class Issue56
     {
         [TestMethod]
         public void SerializeContainsWithNullablesAndWithoutNullableKey()
